Use and instantiate pointers from structs, define types and create/define Factories/Constructors for Interfaces.

+++ pkg/game/GameLoopInterface.go
@@ -1,5 +1,5 @@
 package game
 
 type GameLoopInterface interface {
-	Start(tickRate int, update func())
+	Start(tickRate byte, update func())
 }
